Add application list to root route
