+++ pages/index.tsx
@@ -1,63 +1,77 @@
-import { User } from 'firebase/auth';
-import type { NextPage } from 'next';
-import Head from 'next/head';
-import { useContext, useEffect, useState } from 'react';
-import AuthContext from '../context/AuthContext';
+import { ApplicationStageType, RoleType } from '@prisma/client';
+import { Col, Form, Row, Spin } from 'antd';
+import Search from 'antd/lib/input/Search';
+import Title from 'antd/lib/typography/Title';
+import { ChangeEventHandler, useState } from 'react';
+import useSWR from 'swr';
+import { ApplicationListData, ApplicationQueryParams } from '../../types/application';
+import api from '../api/api';
+import { APPLICATIONS_API_ENDPOINT } from '../api/applicationsApi';
+import ApplicationListCard from '../components/cards/ApplicationListCard';
+import ApplicationStagesSelect from '../components/forms/ApplicationStagesSelect';
+import RoleTypesSelect from '../components/forms/RoleTypesSelect';
+import { ApiResponse } from '../types/apiResponse';
+import { splitByWhitespaces } from '../utils/strings/formatters';
 
-const Home: NextPage = () => {
-  const { currentUser } = useContext(AuthContext);
-  const [jwtToken, setJwtToken] = useState<string>();
+function Applications() {
+  const [searchParams, setSearchParams] = useState<ApplicationQueryParams>({
+    searchWords: [],
+    roleTypeWords: [],
+    stageTypeWords: [],
+  });
 
-  useEffect(() => {
-    currentUser?.getIdToken().then(setJwtToken);
-  }, [currentUser, setJwtToken]);
+  const onSearchBarChange: ChangeEventHandler<HTMLInputElement> = (e) => {
+    setSearchParams({ ...searchParams, searchWords: splitByWhitespaces(e.target.value) });
+  };
 
-  return (
-    <div>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
+  const onRoleTypesFilterChange = (roleTypes: RoleType[]) => {
+    setSearchParams({ ...searchParams, roleTypeWords: roleTypes });
+  };
 
-      <main>
-        <h1 className="text-3xl italic text-rose-500 bg-amber-300 rounded-md p-2">
-          Welcome to <a href="https://nextjs.org">Next.js!</a>
-        </h1>
+  const onApplicationStageTypesFilterChange = (stageTypes: ApplicationStageType[]) => {
+    setSearchParams({ ...searchParams, stageTypeWords: stageTypes });
+  };
 
-        <div className="space-y-4">{displayUserInfo(currentUser)}</div>
-        <div className="space-y-4 p-8">
-          <div className="break-all">{jwtToken || 'JWT token not found'}</div>
-          <p>
-            Check out the JWT token at{' '}
-            <a className="text-sky-500" href="https://jwt.io">
-              jwt.io
-            </a>
-          </p>
-        </div>
-      </main>
-    </div>
+  const { data } = useSWR<ApiResponse<ApplicationListData[]>>(
+    [APPLICATIONS_API_ENDPOINT, searchParams],
+    (url, searchParams) => api.get(url, { params: searchParams }),
   );
-};
 
-function displayUserInfo(user: User | undefined) {
-  if (!user) return <>Failed to get user!</>;
+  const applications: ApplicationListData[] = Array.isArray(data?.payload)
+    ? (data?.payload as ApplicationListData[])
+    : [];
 
-  const githubProviderData = user.providerData.length != 0 ? user.providerData[0] : null;
   return (
-    <>
-      <p>Logged in as UID {user?.uid}</p>
-      {githubProviderData?.photoURL ? (
-        <div className="flex space-x-4">
-          <p>Github Provider Data:</p>
-          <img src={githubProviderData.photoURL} alt="Github Profile Photo" width="30"></img>
-          <p>{githubProviderData.email}</p>
-        </div>
-      ) : (
-        <p>Not signed in to github</p>
-      )}
-    </>
+    <div className="p-5">
+      <Title>My Applications</Title>
+
+      {/* Search and Filters */}
+      <Form>
+        <Row gutter={[16, 16]}>
+          <Col xs={24} md={18}>
+            <Search placeholder="Search..." onChange={onSearchBarChange} />
+          </Col>
+          <Col xs={12} md={3}>
+            <Form.Item>
+              <RoleTypesSelect isMultiselect onChange={onRoleTypesFilterChange} />
+            </Form.Item>
+          </Col>
+          <Col xs={12} md={3}>
+            <Form.Item>
+              <ApplicationStagesSelect isMultiselect onChange={onApplicationStageTypesFilterChange} />
+            </Form.Item>
+          </Col>
+        </Row>
+      </Form>
+
+      {/* Application List */}
+      <Spin spinning={!data}>
+        {applications.map((application, index) => (
+          <ApplicationListCard key={index} application={application} />
+        ))}
+      </Spin>
+    </div>
   );
 }
 
-export default Home;
+export default Applications;
